+++ services/workflows-service/src/events/workflow-completed-webhook-caller.ts
@@ -12,6 +12,7 @@ import { WorkflowRuntimeData } from '@prisma/client';
 import { sign, StateTag } from '@ballerine/common';
 import type { TAuthenticationConfiguration } from '@/customer/types';
 import { CustomerService } from '@/customer/customer.service';
+import { WorkflowRuntimeDataRepository } from '@/workflow/workflow-runtime-data.repository';
 
 @Injectable()
 export class WorkflowCompletedWebhookCaller {
@@ -24,6 +25,7 @@ export class WorkflowCompletedWebhookCaller {
     private readonly logger: AppLoggerService,
     private readonly workflowService: WorkflowService,
     private readonly customerService: CustomerService,
+    private readonly workflowRuntimeDataRepository: WorkflowRuntimeDataRepository,
   ) {
     this.#__axios = this.httpService.axiosRef;
 
@@ -63,7 +65,28 @@ export class WorkflowCompletedWebhookCaller {
       customer.authenticationConfiguration as TAuthenticationConfiguration;
 
     for (const webhook of webhooks) {
-      await this.sendWebhook({ data, webhook, webhookSharedSecret });
+      let childWorkflowsRuntimeData;
+
+      if (webhook.config?.withChildWorkflows) {
+        childWorkflowsRuntimeData = await this.workflowRuntimeDataRepository.findMany(
+          {
+            where: {
+              parentRuntimeDataId: data.runtimeData.id,
+              deletedAt: null,
+            },
+          },
+          [data.runtimeData.projectId],
+        );
+      }
+
+      await this.sendWebhook({
+        data: {
+          ...data,
+          ...(webhook.config?.withChildWorkflows ? { childWorkflowsRuntimeData } : {}),
+        },
+        webhook,
+        webhookSharedSecret,
+      });
     }
   }
 
@@ -80,7 +103,7 @@ export class WorkflowCompletedWebhookCaller {
 
     try {
       // Omit from data properties already sent as part of the webhook payload
-      const { runtimeData, correlationId, entityId, ...restData } = data;
+      const { runtimeData, correlationId, entityId, childWorkflowsRuntimeData, ...restData } = data;
       const {
         createdAt,
         resolvedAt,
@@ -104,6 +127,7 @@ export class WorkflowCompletedWebhookCaller {
         environment,
         data: {
           ...restRuntimeData.context,
+          childWorkflowsRuntimeData,
         },
       };