+++ apps/backoffice-v2/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @ballerine/backoffice-v2
 
+## 0.7.87
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.63
+  - @ballerine/workflow-browser-sdk@0.6.82
+  - @ballerine/workflow-node-sdk@0.6.82
+
+## 0.7.86
+
+### Patch Changes
+
+- @ballerine/workflow-browser-sdk@0.6.81
+- @ballerine/workflow-node-sdk@0.6.81
+
+## 0.7.85
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.61
+  - @ballerine/workflow-browser-sdk@0.6.80
+  - @ballerine/workflow-node-sdk@0.6.80
+
 ## 0.7.84
 
 ### Patch Changes