+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/ui
 
+## 0.5.56
+
+### Patch Changes
+
+- Updated user-facing social media view
+
+## 0.5.55
+
+### Patch Changes
+
+- Fixed phone input styling
+
 ## 0.5.54
 
 ### Patch Changes