+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.6.82
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.63
+  - @ballerine/workflow-core@0.6.82
+
+## 0.6.81
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.81
+
+## 0.6.80
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.80
+  - @ballerine/common@0.9.61
+
 ## 0.6.79
 
 ### Patch Changes